MOve WKImageRecommendationsViewModel to View Model
