+++ Components/Sources/Components/Components/Suggested Edits/Image Recommendations/WKImageRecommendationsViewModel.swift
@@ -51,7 +51,25 @@ public final class WKImageRecommendationsViewModel: ObservableObject {
             self.notSureButtonTitle = notSureButtonTitle
         }
     }
-    
+
+    public struct WKImageRecommendationData {
+        public let pageId: Int
+        public let image: String
+        public let filename: String
+        public let thumbUrl: String
+        public let fullUrl: String
+        public let description: String?
+
+        public init(pageId: Int, image: String, filename: String, thumbUrl: String, fullUrl: String, description: String?) {
+            self.pageId = pageId
+            self.image = image
+            self.filename = filename
+            self.thumbUrl = thumbUrl
+            self.fullUrl = fullUrl
+            self.description = description
+        }
+    }
+
     final class ImageRecommendation: ObservableObject {
         
         let pageId: Int
